services/api/pyproject.toml
[project]
name = "tbg-api"
version = "0.1.0"
description = "The Bitcoin Game — Backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "redis[hiredis]>=5.2.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "pydantic[email]>=2.10.0",
    "structlog>=24.4.0",
    "python-json-logger>=2.0.0",
    "httpx>=0.28.0",
    "PyJWT[crypto]>=2.9.0",
    "argon2-cffi>=23.1.0",
    "coincurve>=20.0.0",
    "arq>=0.26.0",
    "aiosmtplib>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "testcontainers[postgres,redis]>=4.8.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "sqlalchemy[mypy]>=2.0.35",
    "freezegun>=1.4.0",
    "factory-boy>=3.3.0",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "ANN", "ASYNC", "B", "A", "C4", "DTZ", "ICN", "PIE", "PT", "RSE", "RET", "SLF", "SIM", "TID", "TCH", "ARG", "PTH", "ERA", "PL", "TRY", "FLY", "PERF", "FURB", "RUF"]
ignore = ["TRY003"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN", "ARG", "PLR2004", "PT004", "SLF001"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["sqlalchemy.ext.mypy.plugin", "pydantic.mypy"]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["coincurve.*", "aiosmtplib.*", "aioboto3.*", "argon2.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src/tbg --cov-report=term-missing --cov-fail-under=65"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
